backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-pptx==0.6.23
python-multipart==0.0.9
httpx==0.27.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0
python-dotenv==1.0.1
lxml==5.2.2
Pillow==10.3.0
